revert(tournament): drop league-aware edit-scope over-engineering (#274)
Revert the userCard/editUserSchema/hasErrors changes from the league-aware
attempt. They fixed league-tournament MMR editing but regressed nickname editing
on tournament cards (02-tournament-edit): forcing org/league scope for users
without a loaded OrgUser link makes the org/league PATCH throw, and league scope
depends on currentLeague.organization which isn't always populated.

Keeps the safe, regression-free fix (PlayersTab passes organizationId so cards
resolve org scope where currentOrg is loaded). Full league-tournament MMR
editing needs deeper work on currentOrg-loading/orgData-population timing on the
tournament page — tracked separately.
