+++ frontend/app/pages/tournament/tabs/PlayersTab.tsx
@@ -112,11 +112,10 @@ export const PlayersTab: React.FC = memo(() => {
           deleteButtonType="tournament"
           cols={cols}
           emptyMessage="No players in this tournament"
-          // Org/league context so each UserCard resolves a scope-aware edit
-          // (editable MMR + orgUserPk for the PATCH). Without it the edit
-          // pencil falls back to global scope, which strips the MMR field.
+          // Org context so each UserCard resolves orgEntry -> org-scoped edit
+          // (editable MMR). Without it the edit pencil falls back to global
+          // scope and the MMR field is stripped / blocked for org admins.
           organizationId={tournament?.organization_pk ?? currentOrg?.pk}
-          leagueId={tournament?.league_pk ?? undefined}
         />
       </div>
